+++ tools/xdhq/wrappers/JRE/XDHq.java
@@ -26,7 +26,7 @@ public class XDHq extends info.q37.jreq.JREq {
 	static {
 		launcher = JREq.register( "xdhq" );
 	}
-	static protected Object call(int index, Object... objects) {
+	static public Object call(int index, Object... objects) {
 		return JREq.call( launcher, index, objects );
 
 	}
@@ -38,6 +38,6 @@ public class XDHq extends info.q37.jreq.JREq {
 		return (String)call( 0, argument );
 	}
 	static public void launch( String newSessionAction ) {
-		call( 7, "53752", newSessionAction );
+		call( 2, "53752", newSessionAction );
 	}
 }