+++ tools/xdhq/wrappers/prxy_cmn.h
@@ -46,7 +46,7 @@ namespace prxy_cmn {
 		rSetProperty,
 		rGetProperty,
 		rFocus,
-		rNew,
+		rLaunch,	// New command which will replace all the other ones.
 		r_amount,
 		r_Undefined
 	};